+++ packages/api/src/main.ts
@@ -2,6 +2,6 @@ export * from 'apollo-server-lambda'
 export * from './makeServices'
 // @ts-ignore
 export * from './makeMergedSchema/makeMergedSchema'
-export * from './functions/graphQLServer'
+export * from './functions/graphql'
 export * from './globalContext'
 export * from './auth/authHeaders'
